Battle System Update:
An abstract method was implemented to house a character set of skills. Further testing is required, but the abstract method has allowed for the translation of "skill damage formula to skill damage information" to be passed to the main battle system object. Should this method work, it will allow for damage monitoring for any character's skill within the battle. This method can then be further improved to allow enemies to have their own set of skills and damage calculation.