Battle System Update:
Added state control functionality to the battle system script and a function that generate a random target for the enemy to attack. This function will also be able to select which skill to use on the chosen target in later implementation. A new class was added to support this function by transferring the information of who was the attacker, who is the target, and later which skill was used. Character's skills and skill damage formula is still queued for testing.