Battle System Update:
Reworked the skill system where each skill is a child class. Multiple creations of these child skill classes allow for a better customization for each character or enemy. This also allow the implementation of random skill selection by enemies. The battle system is further improved: enemies are now able to attack a character and deals damage to them. The damage dealt is reflected on the character's HP bar and value, and the system maintain this change for the whole battle. Next implementation is creating a death state for characters.