Battle System Update:
Added cooldown system to character skill. The system is implemented by adding a new variable to the base skill class called skillCooldown. Once a character is spawned into the battle, a List<int> will be initialized in HeroStateMachine to keep track of all current skill cooldown of a character within the battle. The battle system is then used to cross check the current cooldown with skillCooldown. If the values don't match, then a character cannot perform the selected skill. Once a turn passes (a turn passes when the enemy has made their move), a function in HeroStateMachine is called to update all current skill cooldown. Also added new skills to the character Melinda, that also account for the cooldown changes.