Battle System Update:
Further enhances the state machine of the battle overseer. The battle overseer now properly separate between the player's turn and the enemy's turn by observing character animation. This required 2 prefabs to have their animation created and added to the system (Alexiel's and Melinda's). The battle overseer now observes the player's animation. As long as the character is still performing their animation, the enemy cannot act. Once all animation is done playing, the battle overseer will give the green light for the enemy to act.